requirements.txt
python-telegram-bot==21.3
openai==1.42.0
langdetect==1.0.9
python-dotenv==1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.6.0
google-auth-httplib2>=0.1.0
requests>=2.28.0
